+++ CHANGES.md
@@ -1,3 +1,9 @@
+#### unreleased changes
+
+* New features:
+* Bug fixes:
+* Under the hood:
+
 #### 2.0.1 / 2026-07-10
 
 * The changelog was accidentally removed from the distribution. It has been restored.